[Fix] Display search result by incrementing ID by default.
Now it makes more sense to sort the songs by release date, roughly reflected in ID.

Also, force an initial sort in frontend to avoid hardcoding in JS. This avoids any assumptions on backend behavior.
One might be tempted to return sorted entries in backend, but searching on multiple fields (soon to be implemented) interferes with this.

Also we bring back the ID display in frontend (makes sense)